Adding clean capability to the Address Cache.
This clean functionality is needed to resolve an issue
of a hostname resolution to different IP addresses when
connected to different networks. This issue is reported
by phone manufacturers and also is described online.

Issue Scenario Description:
Suppose the phone is browsing to a specific page (on a
hostname H) while connected to WiFi. Assume that
hostname H resolves to IP1. When WiFi is not available,
the phone gets connected to cellular data connection.
On the new network, H would resolve to IP2. The issue
is that IP1 stays in DNS cache, which does not get
flushed upon transition to the new network. The result
of IP1 being stored in DNS cache causes the browser
to try to establish connection to IP1 instead of IP2
resulting in connection failure.

(cherry picked from commit 3df62e7025959acc4cc9a813c7698a19f1aee545)

Change-Id: I5e90260888e8fa09635f2e2073223a929b077b46