Avoid a wavy result for InterruptReadable/WritableChannel test
Currently Thread#interrupt() API cannot interrupt (just close fd)
even if a thread is blocking in I/O operation for an
InterruptibleChannel. In other words, if a thread is blocked by
read()/write() operation, then the Thread#interrupt() from another
thread will close fd, but it cannot interrupt the blocking operation
of Channel. In this test case, if the calling sequence is interrupt()
-->read()/write() (actually it occurs on many dual core ICS devices),
then the test passes. Because the fd has already been closed.However,
if the order of the calling sequence would change to read()/write()
--> interrupt(), the test fails (i.e, I/O operation will not return).

Change-Id: Iad1322be101f5b6849678bc505c64ae2ccc78cd1
CRs-Fixed: 382749